Update patch set 3
Patch Set 3:

Yes the loader is instructed by the *lds.S file. The linker is a very capable program in its own right.
 
OK, at this point I am just *guessing*, but I have some experience in this area so it is not just wild speculation.

The boot sequence of *our* use of the MEC goes as follows. The chip is reset and starts running code in the manufacturer created rom inside the chip. This is the code that needs the 8kB stack. This code reads a 'tag' in the spi flash (at a fixed location, 256 bytes from the end of the flash, put there by chip/mec1322/util/pack.py) which points to a 'header' which is digitally signed and gives the start, length and entry address of the image. This rom code validates the signature then transfers the correct number of bytes from flash to ram and jumps to the entry point. After this the rom code is never used, and the main image adjusts the stack pointer to point to a different place (so the 8kB is free to be used for the main image. It would make sense to me to put the traditional "bss" space last in the memory map to avoid potential issues with the original stack conflicting with the RW data being initialized.

Patch-set: 3