Update patch set 1
Patch Set 1:

To further explain your question:

1. RX=TX=1
2. Process RX (RX=0, TX=1)
3. Process TX (RX=TX=0)
4. Read IER (=1 or 3 , no new data has come in yet)
  note: IER is interrupt enable register, 1=RX, 3=TX+RX
5. New data comes in after the read instruction. Sets RX=1.
  5.1 Interrupt pending.
6. Set IER=0. RX=TX=0.
  6.1 Restore IER(=1 or 3). RX=1.
  6.2 Interrupt pending. (6 -> 6.1)
  6.3 Interrupt return.
7. New interrupt fires.
8. RX=1. TX=0
9. But we still call uart_process_input(), which uses LSR register rather than ISR to determine if there's data to read.

The reason that we use LSR register to determine if there's data to read is because TX/RX interrupt status is a value of IIR register and with priority level.

For example: the low nibble of IIR=4 if RX pending, TX pending IIR=2, TX+RX pending IIR=4.
The process may missing in the following case:
1. TX=1, RX=0. (IIR=2)
2. if IIR==4, process RX
3. New data comes in. TX=1, RX=1. (IIR=4)
4. if IIR==2, process TX

Patch-set: 1