add 'backdoor' command framework
This patch introduces a facility which would allow to compile in
callbacks for arbitrary commands.

The callbacks are associated with 16 bit command codes. On input the
callback receives a buffer, containing the command's argument, the
size of the command and the maximum size of the buffer. On output the
callback stores processing result in the same buffer and updates the
size to the actual amount of returned data.

Callback descriptors are stored in a dedicated read only section which
is scanned by backdoor_route_command() to fine a callback associated
with a certain command code.

Typically this will be used for testing, when various test commands
are multiplexed over an existing protocol.

A console channel is also being introduced to allow controlling
console output generated by backdoor commands handlers.

BRANCH=none
BUG=chrome-os-partner:47524
TEST=none yet

Change-Id: I8ae16a78ca7d72176a5e7f74dd7a232078e7c06c
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>