add 'backdoor' command framework
This patch introduces a facility which would allow to register
callbacks for arbitrary commands.

Registered commands expect to receive a sized buffer, containing the
command's argument and a sized buffer to return the result.

This is going to be used for testing, when various test commands are
multiplexed over the existing protocol.

BRANCH=none
BUG=chrome-os-partner:47524
TEST=none yet

Change-Id: I8ae16a78ca7d72176a5e7f74dd7a232078e7c06c
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>