cr50: add backdoor handler for AES testing
This patch registers a handler for AES tests.

The format of the payload the handler expects is as follows:

  field       |    size  |              note
  ================================================================
  mode        |    1     | 0 - decrypt, 1 - encrypt
  cipher_mode |    1     | ECB = 0, CTR = 1, CBC = 2, GCM = 3
  key_len     |    1     | key size in bytes (16, 24 or 32)
  key         | key len  | key to use
  iv_len      |    1     | either 0 or 16
  iv          | 0 or 16  | as defined by iv_len
  text_len    |    2     | size of the text to process, big endian
  text        | text_len | text to encrypt/decrypt

Current limitations are that actual size of the payload should not
exceed 2036 bytes and only ECB cipher mode is supported.

The response generated by the handler consists of the result of the
encrypt/decrypt operation, it is an integer amount of 16 byte blocks.

The test code is compiled in if CRYPTO_TEST_SETUP is defined, which it
presently is.

BRANCH=none
BUG=chrome-os-partner:47524
TEST=Using the python utility on the host, sent some data to the AES
     stack and saw it processes. No actual correctness verification
     yet. The correctness will be verified when the test utility is
     modified to use NIST vector definitions.

Change-Id: I148f286d614f0212192a95b5518cddd32935f43d
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>