Update patch set 4
Patch Set 4:

My suspicion is that the bug you are describing is occurring because the EC and PD firmware versions do not match. There has been a lot of churn around the way the EC and PD interact with each other, and you can run into all sorts of errors if the two are not at the same version number. So, when doing a git bisect to track down a problem, make sure you load both EC and PD with each version. If you are still having a problem, open a bug and we can discuss on the bug.

Patch-set: 4
Reviewer: Gerrit User 1103909 <1103909@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0