Update patch set 1
Patch Set 1:

Here is the problem with the existing commit as I see it -- we had an ectool command that functioned fine on a system without memmap. Now with this CL, on a system without memmap, we'll get error output in addition to the usual output -- so we've made things worse.

If we add a new ectool command then we won't be making things worse for anyone. Code size shouldn't be a problem since it will just inflate the size of ectool slightly, not any code that lives on the EC.

Patch-set: 1