Update patch set 4
Patch Set 4: -Code-Review Verified+1

In this patchset, I added unit tests for the interpreter as well as fixed the bug where the interpreter would lock up on commands that don't produce output. Now, we just assume that the command succeeded until we find an error. If we do, we just retry the last command that we sent.

Patch-set: 4
Label: Verified=+1