Update patch set 3
Patch Set 3:

How about moving it back to your chip-level i2c file and gate it by CONFIG_CMD_I2C_SCAN (not I2C_SCANIO_CMD - ideally it should be a CONFIG, and it's probably not worth adding a new config for one so similar to an existing one)? That is my original suggestion as well.

Patch-set: 3
Label: Code-Review=0