Update patch set 3
Patch Set 3: Code-Review-1

We need the fan to be under EC control until the OS comes up, if it boots to recovery screen it will overheat.

If there is an issue with the fan at high speed early in boot perhaps there is a sensor init timing issue in the EC where it is reading an invalid temperature?

Patch-set: 3
Reviewer: Gerrit User 1000100 <1000100@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1