Update patch set 1
Patch Set 1:

I looked all over in the EC code to find other instances where a global variable is set multiple times in a function and is accessed by both interrupts and tasks and made those accesses volatile. 

LTO is enabled for Glados for some-time now so I think if we have any other related issues e.g. if a global is accessed by multiple tasks, they are probably few and we can fix them as we encounter.

Patch-set: 1