skylake: save/restore wake masks in S0 <-> S0ix transitions
Every 300 seconds or so, the PCH wakes up the EC from S0ix.
In this state transition, it does not wake up the CPU and so
the kernel does not get a call. After handling this, the EC goes back to S0ix.

EC clears the wake mask set by the kernel in the lpc_resume call
before going back to S0ix

In this patch, EC saves and restores the wake masks using two new
hooks for freeze and thaw.

BRANCH=none
BUG=chrome-os-partner:48834
TEST=test lidopen in S0ix

Change-Id: I23751680788ee7a239e321309a1334d37adc4f43
Signed-off-by: Archana Patni <archana.patni@intel.com>
Signed-off-by: Subramony Sesha <subramony.sesha@intel.com>