Update patch set 1
Patch Set 1: Code-Review-1

As Vincent said, this will break every charger that can't supply 3A.

This is actually working as expected. When the system is unlocked, CONFIG_SYSTEM_UNLOCKED, then we allow booting even when battery is < CONFIG_CHARGER_MIN_BAT_PCT_FOR_POWER_ON, because we assume the developer knows what they are doing.

This case is actually similar to the no-battery case, which should actually work as is right now. The reason the no-battery case works is because in get_desired_input_current() if the system is unlocked and there is no battery, we purposely set the initial input current limit to the max and assume the developer must be using a high enough power charger. We could do something similar for your case here and change get_desired_input_current() to also account for a battery that is present but < CONFIG_CHARGER_MIN_BAT_PCT_FOR_POWER_ON.

Patch-set: 1
Label: Code-Review=-1