Update patch set 1
Patch Set 1:

See system_is_locked() in common/system.c. CONFIG_SYSTEM_UNLOCKED overrides the state of the EC to always-unlocked. If you remove the config, but don't bother to assert the WP gpio and lock the RO region, the system will still be unlocked and removing the config won't have achieved anything. To get the system into a locked state, remove the config, assert the WP gpio, and protect the RO region.

Patch-set: 1