Update patch set 1
Patch Set 1:

I probably still have some bugs in this.  It was getting hard to think and it's time for me to take off (and hard to test keyboard changes from the bus), so I uploaded what I have.

Feel free to poke holes in this, or wait until I mark +1 Verified.

Patch-set: 1
Label: Code-Review=0