Update patch set 1
Patch Set 1:

Hi Shawn,

Since Kunimitsu system is unlocked. The factory overwrite always happens. That means even battery is low/charger negotiation is not done, the system is allowed to power up right away. And this causes rebooting loop with some battery.
We need to support "boot up with no battery" case, so we could not just lock the system(and if locked, what are the other impact to development and verification, I'm not sure); meanwhile, we don't want to allow boot up if battery<1% and power<15watt. That's how I add this flag CONFIG_CHARGER_NO_POWER_ON_OVERWRITE, so that if battery is present, no overwrite. And this flag can be removed later on the system is not forced unlocked.

Thanks.

Li

Patch-set: 1