Update patch set 1
Patch Set 1:

If we look at charge_prevent_power_on(), it's being called in two places:

1. Button driver - to skip button press.
2. Power sequencing drivers - to prevent transition out of G3 / S5.

The idea behind allowing power-up when the system is unlocked is -- the user has pushed a button and indicated they want to power up the system. They are responsible for the consequences and should know better if things go wrong (system is unlocked, so this doesn't apply to end-users).

The idea behind requiring the charger to be initialized before allowing power-on is -- in some cases, the user has not indicated they want to power-on the system (they didn't press any button) yet the system powered on anyway (ex. on cold boot). If we're likely to brown out then we should prevent or delay boot _even if the system is unlocked_. That's the function of this line:

prevent_power_on &= (system_is_locked() || charger_is_uninitialized);

If the charger is uninitialized then prevent_power_on is uncleared (regardless of what system_is_locked() returns). This contradicts your CL description a bit.

Is this the exact case you're trying to fix here?

- Low battery
- No charger
- Cold boot (ex. with "reboot" from EC console, or just remove + insert battery)

Result: System goes to S0 and browns out because battery can't supply enough current.

Patch-set: 1