Update patch set 1
Patch Set 1:

Maybe charge_prevent_power_on() should take a parameter "int button_pressed" which indicates whether the power-on was directly caused by pressing a button or not (actually that's not quite what it indicates, since we'll make two checks on our skylake platforms on a single button press -- once in the button driver, once in the power driver). Anyway, we can have a static variable inside charge_prevent_power_on() that will be 1 if a button was ever pressed to power on the device. Then, we can adjust our return value accordingly. What do you think?

Patch-set: 1