Update patch set 1
Patch Set 1:

The main motivation for the force-power-on case is to allow no-battery power-on (which is done sometimes in the factory or by developers).

charge_prevent_power_on() is already difficult to understand (maybe we could write a doc about it). Adding an additional config makes things even more complex (to be fair, so does adding a new function parameter).

What do you think about the following change? Currently, we ignore MIN_BAT_PCT_FOR_POWER_ON iff the system is unlocked AND the charger is initialized. Let's change it to isgnore iff the system is unlocked AND the charger is initialized AND we have a charger attached.

Patch-set: 1