flash_ec: Disconnect EC-3PO interps when flashing.
Due to the recent changes for EC-3PO in servod, flash_ec needs to be
modified in the same way that was done for ToT.

Since the stm32 MCUs are programmed over the UART, we need to make
some changes to allow the interpreter to stop listening to the UART
PTY when flash_ec needs those PTYs.  Otherwise, the EC-3PO interpreter
will interfere with the programming and cause the flash to fail every
time.

BUG=chromium:571170
BUG=chromium:587600
BRANCH=glados
TEST=Use flash_ec to program veyron_jerry with no errors using current
hdctools.

Change-Id: I02746a817fe715b2cb8b58d14d1409b36fb04088
Signed-off-by: Aseda Aboagye <aaboagye@google.com>