flash_ec: Stop EC-3PO interpreters when flashing.
Since the stm32 MCUs are programmed over the UART, we need to make some
changes to allow the interpreter to be stopped when flash_ec needs those
PTYs.  Otherwise, the EC-3PO interpreter will interfere with the
programming and cause the flash to fail every time.

BUG=chromium:571170
BRANCH=None
TEST=Use flash_ec to program both veyron_jerry and samus_pd with no
interruptions.
TEST=Repeat above test but with servod started in a terminal in which
the cros_sdk was entered with --no-ns-pid.

CQ-DEPEND=CL:320439

Change-Id: I350fdb708d30c4ec6f18e5dc4abd621370522381
Signed-off-by: Aseda Aboagye <aaboagye@google.com>