flash_ec: Disconnect EC-3PO interps when flashing.
Due to the recent changes for EC-3PO in servod, flash_ec needs to be
modified in the same way that was done for ToT.

Since the stm32 MCUs are programmed over the UART, we need to make
some changes to allow the interpreter to stop listening to the UART
PTY when flash_ec needs those PTYs.  Otherwise, the EC-3PO interpreter
will interfere with the programming and cause the flash to fail every
time.

BUG=chromium:571170
BUG=chromium:587600
BRANCH=kitty
TEST=Tested same patch on other branches to flash stm32 ECs.

Change-Id: I6f8de0ca5fdc877d2ef7dc613d1f813edaba8187
Signed-off-by: Aseda Aboagye <aaboagye@google.com>