Update patch set 1
Patch Set 1:

Let's revisit Alec's comment above:

This case is actually similar to the no-battery case, which should actually work as is right now. The reason the no-battery case works is because in get_desired_input_current() if the system is unlocked and there is no battery, we purposely set the initial input current limit to the max and assume the developer must be using a high enough power charger. We could do something similar for your case here and change get_desired_input_current() to also account for a battery that is present but < CONFIG_CHARGER_MIN_BAT_PCT_FOR_POWER_ON.

That should prevent us from overcurrenting on boot, and it's simpler than the other proposed changes.

Patch-set: 1