Update patch set 3
Patch Set 3: Verified+1

kunimitsu doesn't pull up the pin on the EC side, but I checked the schematic and kunimitsu does have an external PU (unlike glados), so it makes sense to use OD on the PD side.

I think either an external PU or PU'd on the EC side makes sense, since it's an EC interrupt, and the state of the pin on the PD might be uninitialized if it's rebooting / halting.

Patch-set: 3
Reviewer: Gerrit User 1001696 <1001696@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=+1