cortex-m: fix hard-faults during software interrupt calls
From time to time (usually under heavy interrupt load), the runtime on
Cortex-M0 was panic'ing at the "svc" instruction with a HardFault
exception (inside the wait_event() function).
The issue was probably the following :
the wait_event() code is doing an atomic_read_clear() whose critical
section disables interrupts and re-enables them using "cpsie i",
then do __schedule() call which is essentially a "svc" instruction.
According to ARMv6-m reference manual :
"If execution of a CPS instruction:
increases the execution priority, the CPS execution serializes that
change to the instruction stream.
decreases the execution priority, the architecture guarantees only that
the new priority is visible to instructions executed after either executing
an ISB instruction, or performing an exception entry or exception
return."
So, when we are executing the "svc", PRIMASK.PM can still be seen as 1
(while it was set to 0 by "cpsie i") and in that case the software
interrupt is replaced by a HardFault.

(ported cortex-m0 patch to cortex-m
cortex-m0: fix hard-faults during software interrupt calls
  Ie355c36f06e6fe2fee5cca8998a469fa096badad )

BUG=None
TEST=None

Change-Id: Id6f190a4c614b5b9a0206099509f2a145fc96709
Signed-off-by: Jenny TC <jenny.tc@intel.com>
Signed-off-by: Shobhit Srivastava <shobhit.srivastava@intel.com>