Update patch set 3
Patch Set 3:

batt_str is a pointer inside the memmap data, I don't see how it's relevant at all.

See smbus spec, page 32-33 (Block write-block read process call):
http://smbus.org/specs/smbus20.pdf

"The read byte count does not include the PEC byte. The PEC is computed on the total message beginning
with the first slave address and using the normal PEC computational rules. It is highly recommended that a
PEC byte be used with the Block Write-Block Read Process Call."

We already have a buffer big enough for the maximum data payload (see call to shared_mem_acquire() in smbus_read_block). There's no reason to pass in a larger destination buffer and then only use the first 7 bytes.

We need to restructure the smbus transaction. Rather than having a single call to i2c_xfer, we should complete our write, then read back the byte count. In the next transaction, we should perform a number of reads equal to the read-back byte count + 1 (PEC / CRC). Then, we can calculate CRC on the entire read-back buffer. It doesn't matter what contents we will keep or throw away.

We should fix this properly on TOT once and for all, not carry-over a hack patch from branch to branch. If you need help with it then open a new crbug -- I or some other Googler should be able to get to it this week.

Anyway, thanks for pointing out this problem, it's a big flaw in our smbus code.

Patch-set: 3