GPIO: Simplify initialization
Previously there were only two uses of gpio_config_pins, one was
gpio_config_module, which passed in GPIO_CONFIG_ALL_PORTS (the only
place this is used), the other was the common I2C code when it needs to
return the SDA and SCL lines to their alternate function after unwedging
the bus.

These uses are so different that it doesn't make much sense to keep a
single function for them.  This change adds a gpio_config_pin that is
simpler to use as it just takes a gpio_signal enum to select the GPIO
to configure.  A simplified version of the gpio_config_pins logic that
doesn't have to switch behavior based on the port parameter moved into
the gpio_config_module function.

Signed-off-by: Anton Staaf <robotboy@chromium.org>

BRANCH=None
BUG=None
TEST=make buildall -j

Change-Id: I92bfb0b520b0aa2165655b2ff5076e428c88631f