Skylake: Smart Battery: Fix for boot from cut-off battery
Battery in cut-off mode takes approximately 2 to 3 seconds to initialize.
If the USBC negotiated power drops below the charger threshold power and
if the power is taken from the battery before it is initialized, system
reboots. To eliminate this issue in battery cut-off mode, added code to
make charger as the only power source and prevented AP from booting till
the USBC negotiated power is greater than the charger threshold power
until the battery is initialized.

BRANCH=none
BUG=chrome-os-partner:49224
TEST=Manually tested on Kunimitsu.
     Used console command 'cutoff' to put the battery in cut-off mode.
     Inserted the adopter to wake the system, system doesn't reboot &
     the battery charges.

Change-Id: Ia5a1457506b4bef0b3dd27993e4b60ae64c8f746
Signed-off-by: Vijay Hiremath <vijay.p.hiremath@intel.com>